requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-groq
python-dotenv
PyMuPDF
faiss-cpu
sentence-transformers
langchain-huggingface